Refactored core folder not yet done
