+++ core/CMakeLists.txt
@@ -9,7 +9,7 @@ add_custom_command(
         -DGIT_VERSION_FILE=${GIT_VERSION_FILE}
         -DSOURCE_DIR=${CMAKE_SOURCE_DIR}
         -P ${CMAKE_CURRENT_SOURCE_DIR}/GenerateGitVersion.cmake
-    DEPENDS 
+    DEPENDS
         ${CMAKE_SOURCE_DIR}/.git/HEAD
         ${CMAKE_SOURCE_DIR}/.git/index
     WORKING_DIRECTORY ${CMAKE_SOURCE_DIR}
@@ -20,34 +20,13 @@ add_custom_command(
 # Create a custom target that depends on the generated header
 add_custom_target(git_version_target DEPENDS ${GIT_VERSION_FILE})
 
-add_library(core
-  Core.cpp
-  ROB.cpp
-  MMU.cpp
-  Preloader.cpp
-  CPU.cpp
-  CPUFactory.cpp
-  CPUTopology.cpp
-)
-
-add_library(instgen
-  Inst.cpp
-  InstArchInfo.cpp
-  InstGroup.cpp
-  InstGenerator.cpp
-)
-
-find_package(Boost REQUIRED COMPONENTS json)
-
-target_link_libraries(instgen ${STF_LINK_LIBS} mavis Boost::json)
-
-get_property(SPARTA_INCLUDE_PROP TARGET SPARTA::sparta PROPERTY INTERFACE_INCLUDE_DIRECTORIES)
-target_include_directories(core SYSTEM PRIVATE ${SPARTA_INCLUDE_PROP})
-
-# Make core library depend on git_version_target
-add_dependencies(core git_version_target)
-target_include_directories(core PUBLIC "${CMAKE_CURRENT_BINARY_DIR}")
-
+add_subdirectory(cpu)
+add_subdirectory(inst)
+add_subdirectory(memory)
+add_subdirectory(rob)
+add_subdirectory(preloader)
+add_subdirectory(utils)
+add_subdirectory(sim_engine)
 add_subdirectory(fetch)
 add_subdirectory(decode)
 add_subdirectory(rename)
@@ -56,4 +35,5 @@ add_subdirectory(execute)
 add_subdirectory(vector)
 add_subdirectory(lsu)
 
-target_link_libraries(core fetch decode rename dispatch execute vector lsu instgen)
+target_include_directories(core_sim_engine PUBLIC "${CMAKE_CURRENT_BINARY_DIR}")
+add_dependencies(core_sim_engine git_version_target)
