+++ core/cpu/CPUTopology.cpp
