+++ core/cpu/CPUTopology.hpp
