+++ core/decode/CMakeLists.txt
@@ -3,4 +3,7 @@ add_library(decode
   Decode.cpp
   MavisUnit.cpp
 )
-target_link_libraries(decode instgen)
+target_include_directories(decode PUBLIC
+    $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/..>
+)
+target_link_libraries(decode PUBLIC core_inst core_cpu)
