+++ core/execute/CMakeLists.txt
@@ -3,3 +3,4 @@ add_library(execute
   ExecutePipe.cpp
   IssueQueue.cpp
 )
+target_link_libraries(execute PUBLIC core_inst core_cpu)
