+++ core/inst/InstArchInfo.cpp
