+++ core/inst/InstGenerator.cpp
