+++ core/inst/InstGenerator.hpp
