+++ core/inst/LoadStoreInstInfo.hpp
