+++ core/memory/CacheFuncModel.hpp
