+++ core/memory/MSHREntryInfo.hpp
