+++ core/memory/MemoryAccessInfo.hpp
