+++ core/memory/ReplacementFactory.hpp
