+++ core/rename/CMakeLists.txt
@@ -1,4 +1,4 @@
 add_library(rename
     Rename.cpp
 )
-target_link_libraries(rename instgen)
+target_link_libraries(rename PUBLIC core_inst core_cpu)
