+++ core/sim_engine/CMakeLists.txt
@@ -1,0 +1,13 @@
+add_library(core_sim_engine STATIC
+    Core.cpp
+)
+
+target_sources(core_sim_engine PUBLIC
+    $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/Core.hpp>
+    $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/SimulationConfiguration.hpp>
+)
+
+target_include_directories(core_sim_engine PUBLIC
+    $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}>
+    $<INSTALL_INTERFACE:include>
+)
