+++ core/utils/CoreExtensions.hpp
