+++ core/utils/CoreUtils.hpp
