+++ core/utils/MiscUtils.hpp
