+++ core/vector/CMakeLists.txt
@@ -1,3 +1,4 @@
 add_library(vector
   VectorUopGenerator.cpp
 )
+target_link_libraries(vector PUBLIC core_inst core_cpu)
