+++ mss/CMakeLists.txt
@@ -6,3 +6,4 @@ add_library(mss
   )
 get_property(SPARTA_INCLUDE_PROP TARGET SPARTA::sparta PROPERTY INTERFACE_INCLUDE_DIRECTORIES)
 target_include_directories(mss SYSTEM PRIVATE ${SPARTA_INCLUDE_PROP})
+target_link_libraries(mss PUBLIC core_memory core_cpu)
