+++ test/core/rename/CMakeLists.txt
@@ -1,11 +1,7 @@
-project(Rename_test)
-
-target_link_libraries (core mss)
-
 add_executable(Rename_test Rename_test.cpp ${SIM_BASE}/sim/OlympiaSim.cpp)
 #add_executable(Rename_test_full Rename_test_full.cpp ${SIM_BASE}/sim/OlympiaSim.cpp)
 
-target_link_libraries(Rename_test core common_test ${STF_LINK_LIBS} mavis SPARTA::sparta)
+target_link_libraries(Rename_test rename core_cpu core_rob core_sim_engine core_utils mss common_test ${STF_LINK_LIBS} mavis SPARTA::sparta)
 
 file(CREATE_LINK ${SIM_BASE}/mavis/json ${CMAKE_CURRENT_BINARY_DIR}/mavis_isa_files SYMBOLIC)
 file(CREATE_LINK ${SIM_BASE}/arches     ${CMAKE_CURRENT_BINARY_DIR}/arches          SYMBOLIC)
